requirements.txt
asgiref==3.5.2
Django==4.1.2
django-cors-headers==3.13.0
django-environ==0.9.0
django-js-asset==2.0.0
django-mptt==0.14.0
django-phonenumber-field==7.0.0
django-versatileimagefield==2.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.1
gunicorn==20.1.0
numpy==1.23.1
opencv-python==4.6.0.66
phonenumbers==8.12.57
Pillow==9.2.0
psycopg2==2.9.3
psycopg2-binary==2.9.4
PyJWT==2.5.0
python-magic==0.4.27
pytz==2022.1
sqlparse==0.4.2
whitenoise==6.2.0
